Move Task and friends to cocotb.task
