+++ documentation/source/library_reference.rst
@@ -71,7 +71,7 @@ Task Management
 
 .. autofunction:: cocotb.create_task
 
-.. autoclass:: cocotb.decorators.Task
+.. autoclass:: cocotb.task.Task
     :members:
 
 Handle Values
